Adding makergear config with rewritten / more clear history
The makergear firmware was a hardfork which made it hard to compare
diffs on github.

This branch is all the makergear changes to marlin 1 in a single diff.

Note that this process was not automatic, and might have errors. Be
vigilant!

To produce this branch I:
1. Edited git history to remove diffs that took out marlin files.
   That is, a few commits removed the example configuration files, I added
   those back in by rebasing.

   # find commits with file deletions
   git log --diff-filter=D --stat

   # edit the commit
   git rebase -Xtheirs -i <sha>~1

   # add files back in and roll forward
   git show --no-renames --name-only --format= --diff-filter=D -z | tr \\n \\0 | xargs -0 git checkout HEAD~1 --
   git commit --amend --no-edit --allow-empty
   git rebase --continue

2. Then I started with an upstream branch of marlin 1.1.6 (since this
   was the last version of marlin the makergear branch merged in)

3. I copied the cleaned up makergear files over and then did a little
   cleanup around file permission changes.

So, this branch _should_ show intended changes to marlin. The goal is
that we can 1-1 map these changes to marlin 2.